pyproject.toml
[project]
name = "cubbi"
version = "0.1.0"
description = "Cubbi Container Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "docker>=7.0.0",
    "pyyaml>=6.0.1",
    "rich>=13.6.0",
    "pydantic>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.9",
    "mypy>=1.7.0",
]

[project.scripts]
cubbi = "cubbi.cli:app"
cubbix = "cubbi.cli:session_create_entry_point"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
